kernels: rename `utils` module to `load`
