+++ kernels/src/kernels/cli/benchmark.py
@@ -20,9 +20,9 @@ from huggingface_hub.utils import (
     hf_raise_for_status,
 )
 
+from kernels.backends import _backend
 from kernels.benchmark import Benchmark
 from kernels.hf_hub import _get_hf_api
-from kernels.utils import _backend
 
 MISSING_DEPS: list[str] = []
