+++ kernels/tests/test_loaded_kernels.py
@@ -4,7 +4,7 @@ import pytest
 
 from kernels import get_kernel, get_loaded_kernels, get_local_kernel, install_kernel
 from kernels.hf_hub import RepoInfo
-from kernels.utils import LoadedKernel, _loaded_kernels
+from kernels.importer import LoadedKernel, _loaded_kernels
 
 _REPO_ID = "kernels-community/relu"
 _PACKAGE_NAME = "relu"
